requirements.txt
python-dotenv
tweepy
halo
